requirements.txt
setuptools==59.1.1
boto3==1.21.12
celery==5.2.3
dj-database-url==0.5.0
Django==3.2.12
django-celery-beat==2.2.1
django-redis==5.2.0
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
drf-yasg==1.20.0
flower==1.0.0
gunicorn==20.1.0
mixer==7.2.1
mock==4.0.3
Pillow==9.0.1
psycopg2==2.9.3
pytest==7.0.1
pytest-django==4.5.2
redis==4.1.4
sentry-sdk==1.5.6
whitenoise==6.0.0
